moving StripeNetworkUtils into the main package to reduce api surface
