+++ stripe/src/main/java/com/stripe/android/StripeApiHandler.java
@@ -22,7 +22,6 @@ import com.stripe.android.model.Token;
 import com.stripe.android.net.StripeResponse;
 import com.stripe.android.net.StripeSSLSocketFactory;
 import com.stripe.android.util.LoggingUtils;
-import com.stripe.android.util.StripeNetworkUtils;
 import com.stripe.android.util.StripeTextUtils;
 
 import org.json.JSONArray;
