+++ stripe/src/main/java/com/stripe/android/model/SourceRedirect.java
@@ -14,7 +14,7 @@ import java.util.Map;
 
 import static com.stripe.android.model.StripeJsonUtils.optString;
 import static com.stripe.android.model.StripeJsonUtils.putStringIfNotNull;
-import static com.stripe.android.util.StripeNetworkUtils.removeNullAndEmptyParams;
+import static com.stripe.android.StripeNetworkUtils.removeNullAndEmptyParams;
 
 /**
  * Model for a <a href="https://stripe.com/docs/api#source_object-redirect">redirect</a> object
