+++ stripe/src/main/java/com/stripe/android/model/SourceSepaDebitData.java
@@ -4,7 +4,7 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.VisibleForTesting;
 
-import com.stripe.android.util.StripeNetworkUtils;
+import com.stripe.android.StripeNetworkUtils;
 
 import org.json.JSONException;
 import org.json.JSONObject;
