+++ stripe/src/test/java/com/stripe/android/StripeNetworkUtilsTest.java
@@ -1,4 +1,4 @@
-package com.stripe.android.util;
+package com.stripe.android;
 
 import android.support.annotation.NonNull;
